[FLINK-17654][core] Move Clock classes to flink-core to make them usable outside runtime
We want to use them in the new WatermarkGenerator implementations in
follow-up commits.
