+++ flink-core/src/main/java/org/apache/flink/util/clock/ManualClock.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.util.clock;
+package org.apache.flink.util.clock;
 
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
